fix regression during size_t conversion

+++ src/tokenizer/io.cc
@@ -43,7 +43,7 @@ void Tokenizer::prevChar() {
 			// count the amount of characters on this line
 			// TODO does this affect parsing performance significantly?
 			int count = 0;
-			for(size_t i = this->fileIndex - 1; i >= 0; i--) {
+			for(int i = this->fileIndex - 1; i >= 0; i--) {
 				if(this->contents[i] == '\n') {
 					break;
 				}
